Split ssh state tests into separate modules to respect naming convention
